Changed Hash Algorithm

+++ 1/hash.h
@@ -65,7 +65,7 @@ class hashTable {
   std::vector<hashItem> data; // The actual entries are here.
 
   // The hash function.
-  unsigned int hash(const std::string &key);
+  unsigned long hash(const std::string &key);
 
   // Search for an item with the specified key.
   // Return the position if found, -1 otherwise.
